pyproject.toml
[tool.poetry]
name = "swanti"
version = "0.1.0"
description = "TSP SWAN script (Experimental)"
authors = ["USoVietnam"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Pillow = "^10.1.0"
pyimgur = "^0.6.0"
pygraphviz = "^1.11"
networkx = "^3.2.1"
requests = "^2.31.0"
google-api-python-client = "^2.108.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
